+++ resource/csdk/stack/samples/linux/SimpleClientServer/ocserverbasicops.h
@@ -36,7 +36,7 @@
 typedef struct LEDRESOURCE{
     OCResourceHandle handle;
     bool state;
-    int power;
+    int64_t power;
 } LEDResource;
 
 //-----------------------------------------------------------------------------
@@ -46,7 +46,7 @@ typedef struct LEDRESOURCE{
 /* Function that creates a new LED resource by calling the
  * OCCreateResource() method.
  */
-int createLEDResource (char *uri, LEDResource *ledResource, bool resourceState, int resourcePower);
+int createLEDResource (char *uri, LEDResource *ledResource, bool resourceState, int64_t resourcePower);
 
 /* This method constructs a response from the request */
 OCRepPayload* constructResponse (OCEntityHandlerRequest *ehRequest);
